Accessing my database and displaying it via a view is working
Finally. Flask is successfully (it appears) accessing my database, and
displaying info via a view & template.

Next step: See if I can get WTForms working.
