+++ templates/index.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
 
 {% block content %}
-		<h1>Hello, {{ user.nickname }}!</h1>
+		<h1>Hello, {{ student.name }}!</h1>
 {% endblock %}
