+++ views.py
@@ -1,8 +1,11 @@
 from flask import render_template
 from app import app
+from data_model import *
+
 
 @app.route("/")
 @app.route("/index")
 def index():
-	user = {"nickname": "Andrew"}
-	return render_template("index.html", title = "Home", user = user)
+	student = Student.select().where(Student.name == "Jonathan Turpen").get()
+
+	return render_template("index.html", title = "Home", student = student)
